ssh file
This 'ssh' file is a blank file, with no file extension. It is placed in the root of the 'boot' partition after burning a new Raspbian image to a Micro SD card, before inserting it in to a Raspberry Pi for the first boot.

It enables SSH on the Raspberry Pi.