wpa_supplicant.conf file
This is template for the wpa_supplicant.conf file, It is placed in the root of the 'boot' partition after burning a new Raspbian image to a Micro SD card, before inserting it in to a Raspberry Pi for the first boot.

You should edit the file to have your 'SSID' (WiFi network name) and 'PSK' (pre-shared key/password). scan-ssid allows it to find "hidden" wifi networks, ones which aren't actively broadcasting their SSID, despite being an active WiFi network. These settings do assume you are using a secured WiFi network with a password (eg: secured with WPA2 or may also be called WPA2-PSK in your router)

If you are not in the United Kingdom you should also change the country from GB to your relevant country code.